# backend/user-authentication/auth_swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_createUser",
        "summary": "User Signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The record found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_getUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "Customer Loyalty backend API",
    "description": "Authentication APIS",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "components": {
    "schemas": {
      "UserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "dob": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "dob",
          "contact"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "DOB": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "email",
          "role",
          "contact",
          "DOB",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "contact": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "contact",
          "password"
        ]
      }
    }
  }
}